refactor(gpu_ntt): split the monolithic ntt.rs into per-concern submodules
Dissolve src/ntt/ntt.rs into sibling submodules under src/ntt/, dropping the
inner `ntt` module name (which triggered clippy's module_inception):

- kernels.rs   — the five cuda_kernel! signature families + registrations
- inverse.rs   — evals_to_monomials_{2,3}_pass
- forward.rs   — the six monomials_to_evals_* launchers
- lde.rs       — the LDE grid/config helpers + multi-coset forward entries
- dispatch.rs  — dispatch_strategy + natural_evals_to_bitreversed_monomials

Structure-only: launcher bodies move byte-identical. The crate-external facade
(gpu_ntt::ntt::*) is unchanged — mod.rs re-exports the same names from the new
submodules; in-crate/test paths get mechanical `super::ntt::X` -> `super::<mod>::X`
updates.

Visibility plumbing (minimum to compile across the new module boundary):
- kernels.rs emits its Signature/Arguments via cuda_kernel_signature_and_arguments!
  (pub(super) ...) and hand-writes each *Function struct with a pub(super) field,
  because the multi-arm cuda_kernel! family form forces module-private items and a
  private tuple field cannot be constructed from a sibling module.
- dispatch_strategy widened private -> pub(super) (consumed by lde.rs).
No pub items demoted; visibility demotions are deferred to a later task.
